+++ lib/modules/ui/ui_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: ui
-// Updated at: 2026-05-04 20:37:21
-// Commit: fix(ui): prevent double tap duplicate action in ui
+// Updated at: 2026-05-04 21:29:15
+// Commit: style(ui): harmonize light and dark theme palette for ui
 
 class uiComponentState {
-  final String commitRef = '859';
-  final String lastMessage = 'fix(ui): prevent double tap duplicate action in ui';
-  final DateTime timestamp = DateTime.parse('2026-05-04T20:37:21+05:00');
+  final String commitRef = '861';
+  final String lastMessage = 'style(ui): harmonize light and dark theme palette for ui';
+  final DateTime timestamp = DateTime.parse('2026-05-04T21:29:15+05:00');
 }